Resolve "Create 2 methods for 1C. Customers and Counters"

Closes #6

See merge request uk.our.home/our.home!6

+++ database/migrations/2024_09_01_180222_create_clients_table.php
@@ -12,7 +12,7 @@ return new class extends Migration {
             $table->string('name');
             $table->string('phone')->unique();
             $table->foreignId('user_id')->nullable()->constrained('users');
-            $table->decimal('debt', 10, 2)->nullable();
+            $table->float('debt', 10, 2)->nullable();
             $table->timestamps();
         });
     }
